recommendation-engine/requirements.txt
fastapi
uvicorn[standard]
httpx
numpy
scikit-learn
pydantic
python-dotenv
